Refactor to collect the many supporting Mat classes into a common package
