+++ visual/src/main/scala/i4s/scalacv/core/model/mats/Matable.scala
@@ -1,4 +1,4 @@
-package i4s.scalacv.core.model
+package i4s.scalacv.core.model.mats
 
 import i4s.scalacv.core.types.Types.Type
