+++ visual/src/main/scala/i4s/scalacv/image/Image.scala
@@ -1,8 +1,8 @@
 package i4s.scalacv.image
 
 import i4s.scalacv.core.constants.AccessFlags.AccessFlag
-import i4s.scalacv.core.model.syntax._
-import i4s.scalacv.core.model.{Mat, Scalar, Size}
+import i4s.scalacv.core.model.mats._
+import i4s.scalacv.core.model.{Scalar, Size}
 import i4s.scalacv.core.types.Types
 import i4s.scalacv.core.types.Types.{Cv8U, Type}
 import org.bytedeco.javacv.{Java2DFrameConverter, OpenCVFrameConverter}
